fix: imageSrc query parameter is UUID for uploaded images
For large uploaded images, putting the base64 representation (which is
the 'src' argument of the image) as a query argument value generates an
URL which length is not manageable by browsers.

Instead, these locally uploaded images are assigned an UUID identifier.

Obviously, sharing such an URL does not carry the image itself anymore,
and another solution will be needed to store the image before sharing.
See https://github.com/LyonDataViz/occupapp/issues/52.