fix: improve state update flow
- remove redundancy in the state (see below)
- cut the parsing and routing processes into smallest pieces (src/utils/...)
- fix a bug that broke the selection of an uploaded image
- reduce a bit the size of the query parameters (imageSrc -> img, etc.).
But since the UUID of all the objects are still in the URL, it's not
that useful for now.

---

The Vuex state is managed as follows.

The URL query parameters are the source of trust, and define the current
composition. Note that if the URL doesn't contain any query parameters, or if
they are insufficient to recreate a complete composition, the application will
be routed to a new URL, with missing parameters filled with default values,
possibly several times, until converging to a valid URL.

Once the URL has been parsed successfully, the current composition is stored in
store/compositions, that contains the compositions associated to all the images
(the default ones, and any other updated or linked image). Note that these
compositions are not persisted in the URL, except the current one.

The Vuex state also contains other data that is not persisted in the URL:

- derived data: (such as store/current/backgroundImage.ts). They don't contain
  any data by their own, but react to current composition change by updating
  their cached getters. They also provide helpers to update the current
  composition (ie. to route to a new URL). They have access to the router.
- temporary data, such as the points selection used to delete various points at
  once (it's cleared every time a composition is changed)
- cached data: store/current/backgroundImage.ts contains the background image,
  and store/current/pointMetrics contains the points metrics that are computed
  from the current composition
- general data, such as the color vs black and white switch for background image