feat: change the data model, no more based on the URL
The URL is no more updated at every state modification.
It is parsed and a composition can be passed through the URL.
There is way currently to share an URL
The compositions have an id, and it's the basis to select the current
composition.
A composition is no more indexed by its background image. Various
compositions can use the same background image.
Locally uploaded images can be shared through URL, with a very reduced
size so that the base64 representation doesn't pass the max URL length
(~2000 is the common accepted limit for URLs). Note that nothing ensures
that the URL will be under 2000 characters.